+++ bluepysnap/cli.py
@@ -1,9 +1,12 @@
 """The project's command line launcher."""
+import functools
 import logging
+import warnings
 
 import click
 
 from bluepysnap import circuit_validation, simulation_validation
+from bluepysnap.utils import Deprecate
 
 CLICK_EXISTING_FILE = click.Path(exists=True, file_okay=True, dir_okay=False)
 
@@ -20,18 +23,50 @@ def cli(verbose):
     )
 
 
+def circuit_validation_params(func):
+    """Small helper to have shared params."""
+
+    @click.argument("config_file", type=CLICK_EXISTING_FILE)
+    @click.option(
+        "--skip-slow/--no-skip-slow",
+        default=True,
+        help=(
+            "Skip slow checks; checking all edges refer to existing node ids, "
+            "edge indices are correct, etc"
+        ),
+    )
+    @click.option("--only-errors", is_flag=True, help="Only print fatal errors (ignore warnings)")
+    @functools.wraps(func)
+    def wrapper(*args, **kwargs):
+        return func(*args, **kwargs)
+
+    return wrapper
+
+
 @cli.command()
-@click.argument("config_file", type=CLICK_EXISTING_FILE)
-@click.option(
-    "--skip-slow/--no-skip-slow",
-    default=True,
-    help=(
-        "Skip slow checks; checking all edges refer to existing node ids, "
-        "edge indices are correct, etc"
-    ),
-)
-@click.option("--only-errors", is_flag=True, help="Only print fatal errors (ignore warnings)")
+@circuit_validation_params
 def validate(config_file, skip_slow, only_errors):
+    """[DEPRECATED] Validate Sonata circuit based on config file.
+
+    Args:
+        config_file (str): path to Sonata circuit config file
+        skip_slow (bool): skip slow tests
+        only_errors (bool): only print fatal errors
+    """
+    with warnings.catch_warnings():
+        # Making sure the warning is shown
+        warnings.simplefilter("always", DeprecationWarning)
+        Deprecate.warn(
+            "Calling circuit validation with 'validate' is deprecated. "
+            "Please use 'validate-circuit' instead."
+        )
+
+    circuit_validation.validate(config_file, skip_slow, only_errors)
+
+
+@cli.command()
+@circuit_validation_params
+def validate_circuit(config_file, skip_slow, only_errors):
     """Validate Sonata circuit based on config file.
 
     Args: